requirements.txt
google-cloud-aiplatform
pandas
pydantic
python-dotenv
sentence_transformers
